backend actualizado

+++ src/main/java/com/proyecto/MarketDelivery/repository/CatalogoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CatalogoRepository extends JpaRepository<Catalogo, Integer> {
+public interface CatalogoRepository extends JpaRepository<Catalogo, Long> {
 }
