+++ src/main/java/com/proyecto/MarketDelivery/repository/EmprendimientoRepository.java
@@ -5,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmprendimientoRepository extends JpaRepository<Emprendimiento, Integer> {
+public interface EmprendimientoRepository extends JpaRepository<Emprendimiento, Long> {
 }
-
